+++ src/templates/hub.config.js
@@ -1,5 +1,18 @@
 module.exports = {
   hasAsteroid: true, // usado caso tenha extensão @bildvitta/asteroid instalada na aplicação
   forbiddenRouteName: 'Forbidden', // usado para definir o nome da rota da pagina de erro 404 (Forbidden)
-  storeAdapter: 'vuex' // usado para definir qual store esta utilizando na aplicação "vuex" ou "pinia"
+  storeAdapter: 'vuex', // usado para definir qual store esta utilizando na aplicação "vuex" ou "pinia"
+  development: { // ambiente de desenvolvimento
+    localhost: {
+      useAutomaticLogin: true, // usado para definir se o login automático sera feito
+      environment: 'development', // flag para definir o ambiente
+      url: '' // url do ambiente a ser acessado
+    },
+
+    preview: {
+      useAutomaticLogin: true, // usado para definir se o login automático sera feito
+      environment: 'temporary', // flag para definir o ambiente
+      url: '' // url do ambiente a ser acessado
+    }
+  }
 }